test: Spring Data Jpa 적용에 따른 전체 테스트 코드 업데이트

+++ src/main/java/roomescape/service/ReservationWaitingService.java
@@ -62,6 +62,9 @@ public class ReservationWaitingService {
 
     @Transactional
     public void delete(Long reservationWaitingId) {
+        if (!reservationWaitingRepository.existsById(reservationWaitingId)) {
+            throw new RoomEscapeException(ReservationWaitingErrorCode.RESERVATION_WAITING_NOT_FOUND);
+        }
         reservationWaitingRepository.deleteById(reservationWaitingId);
     }
